Move Frame classes into their own file
